7.6.8 - Consistency corrections
Some variable types were inconsistent in parts of the libraries; a general fix was made for this. Issues with values ​​being truncated and converted inconsistently were also addressed.

+++ src/lib/ADXL345/ADXL345.h
@@ -10,7 +10,7 @@
 class ADXL345 : public Accel
 {
 public:
-ADXL345(long recalT = 0.1);
+ADXL345(float recalT = 0.1f);
 	void begin();
 	bool readAll();
 };
