+++ src/lib/Sens/Sens.cpp
@@ -1,7 +1,7 @@
 #include "Sens.h"
 
 /// Sensor base class
-Sens::Sens(uint8_t add, long recalT) : address(add), recalibrateT((long)(recalT * 1000000))
+Sens::Sens(uint8_t add, float recalT) : address(add), recalibrateT((long)(recalT * 1000000))
 {
 }
 long Sens::getTimeLapse()
